Add dam to com1DFAorig
  - Getting Walllines to work from cint
  - Reproduce samos results for dam
