+++ avaframe/ana1Tests/energyLineTestCfg.ini
@@ -93,3 +93,5 @@ frictModel = Coulomb
 #+++++++++++++General Friction parameters
 # tan of bed friction angle used for: samosAT, Coulomb, Voellmy
 mu = 0.4
+
+dam = True
