+++ avaframe/com1DFA/com1DFACfg.ini
@@ -301,12 +301,12 @@ entDefResistance = 0
 # the dam foot print is given in Intputs/DAM as a shape file line
 # the dam is located on the left part of the foot print
 # use dam?
-dam = False
-# dam height [m] (vertical)
-damHeight = 5
-# dam slope [°] measured from the horizontal plane
-damSlope = 30
-restitutionCoefficient =
+dam = True
+# behavior of the particles when they hit the dam (are they fully reflected: restitutionCoefficient = 1 or only reflected
+# along the tangent component of the dam: restitutionCoefficient = 0 or something in between: restitutionCoefficient = 0-1)
+restitutionCoefficient = 0
+# 1 to activate energy loss due to snow flowing over dam
+dissDam = 0
 
 
 #++++++++++++++ Technical values +++++++++++++
@@ -339,7 +339,7 @@ createRangeTimeDiagram = False
 # if False for range-time diagram with radar field of view
 TTdiagram = True
 # scenario name - can be used for plotting
-scenario = 
+scenario =
 
 
 [INPUT]
