+++ avaframe/data/avaBowl/Inputs/DAM/dam.cpg
@@ -1,0 +1,1 @@
+UTF-8
