+++ avaframe/data/avaSlide/Inputs/slideTopo.asc
@@ -1,7 +1,7 @@
 ncols     419
 nrows    201
-xllcorner 1.0
-yllcorner 1.0
+xllcenter 0
+yllcenter 0
 cellsize 5
 NODATA_value -9999
 1752.60 1749.10 1745.60 1742.10 1738.60 1735.10 1731.60 1728.10 1724.60 1721.10 1717.60 1714.10 1710.60 1707.10 1703.60 1700.10 1696.60 1693.10 1689.60 1686.10 1682.60 1679.10 1675.60 1672.10 1668.60 1665.10 1661.60 1658.10 1654.60 1651.10 1647.60 1644.10 1640.60 1637.10 1633.60 1630.10 1626.60 1623.10 1619.60 1616.10 1612.60 1609.10 1605.60 1602.10 1598.60 1595.10 1591.60 1588.10 1584.60 1581.10 1577.60 1574.10 1570.60 1567.10 1563.60 1560.10 1556.60 1553.10 1549.60 1546.10 1542.60 1539.10 1535.60 1532.10 1528.60 1525.10 1521.60 1518.10 1514.60 1511.10 1507.60 1504.10 1500.60 1497.10 1493.60 1490.10 1486.60 1483.10 1479.60 1476.10 1472.60 1469.10 1465.60 1462.10 1458.60 1455.10 1451.60 1448.10 1444.60 1441.10 1437.60 1434.10 1430.60 1427.10 1423.60 1420.10 1416.60 1413.10 1409.60 1406.10 1402.60 1399.10 1395.60 1392.10 1388.60 1385.10 1381.60 1378.10 1374.60 1371.10 1367.60 1364.10 1360.60 1357.10 1353.60 1350.10 1346.60 1343.10 1339.60 1336.10 1332.60 1329.10 1325.60 1322.10 1318.60 1315.10 1311.60 1308.10 1304.60 1301.10 1297.60 1294.10 1290.60 1287.10 1283.60 1280.10 1276.60 1273.10 1269.60 1266.10 1262.60 1259.10 1255.60 1252.10 1248.60 1245.10 1241.60 1238.10 1234.60 1231.10 1227.60 1224.10 1220.60 1217.10 1213.60 1210.10 1206.60 1203.10 1199.60 1196.10 1192.60 1189.14 1185.71 1182.32 1178.97 1175.66 1172.38 1169.14 1165.94 1162.77 1159.64 1156.54 1153.48 1150.45 1147.46 1144.50 1141.58 1138.69 1135.84 1133.02 1130.24 1127.49 1124.77 1122.09 1119.44 1116.82 1114.23 1111.68 1109.16 1106.67 1104.21 1101.78 1099.39 1097.03 1094.70 1092.40 1090.13 1087.89 1085.68 1083.50 1081.35 1079.23 1077.14 1075.08 1073.05 1071.05 1069.08 1067.14 1065.23 1063.35 1061.50 1059.68 1057.89 1056.13 1054.40 1052.70 1051.02 1049.37 1047.75 1046.16 1044.60 1043.07 1041.56 1040.08 1038.63 1037.21 1035.82 1034.45 1033.11 1031.80 1030.52 1029.26 1028.03 1026.83 1025.65 1024.50 1023.38 1022.28 1021.21 1020.17 1019.16 1018.17 1017.21 1016.28 1015.37 1014.49 1013.63 1012.80 1012.00 1011.22 1010.47 1009.75 1009.05 1008.38 1007.73 1007.11 1006.52 1005.95 1005.41 1004.89 1004.40 1003.94 1003.50 1003.09 1002.70 1002.34 1002.01 1001.70 1001.42 1001.16 1000.93 1000.72 1000.54 1000.39 1000.26 1000.16 1000.08 1000.03 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00 1000.00
