+++ avaframe/out3Plot/outDebugPlots.py
@@ -2,7 +2,7 @@ import numpy as np
 import pathlib
 import copy
 import matplotlib
-# matplotlib.use('agg')
+matplotlib.use('agg')
 import matplotlib.pyplot as plt
 import matplotlib as mpl
 from mpl_toolkits.axes_grid1 import make_axes_locatable
