+++ avaframe/tests/data/testEnergyLine/energyLineTestCfg.ini
@@ -118,3 +118,5 @@ frictModel = Coulomb
 #+++++++++++++General Friction parameters
 # tan of bed friction angle used for: samosAT, Coulomb, Voellmy
 mu = 0.4
+
+dam = False
