+++ avaframe/tests/test_ana1Tests.py
@@ -196,6 +196,7 @@ def test_mainEnegyLineTest(tmp_path):
     com1DFACfg = cfgUtils.getModuleConfig(com1DFA, fileOverride='', modInfo=False, toPrint=False,
                                           onlyDefault=energyLineTestCfg['com1DFA_override']['defaultConfig'])
     com1DFACfg, energyLineTestCfg = cfgHandling.applyCfgOverride(com1DFACfg, energyLineTestCfg, com1DFA, addModValues=False)
+    print(energyLineTestCfg['com1DFA_override']['dam'], com1DFACfg['GENERAL']['dam'])
     com1DFACfgFile = cfgUtils.writeCfgFile(avalancheDir, com1DFA, com1DFACfg, fileName='com1DFA_settings', filePath=workPath)
     dem, simDF, _ = com1DFA.runOrLoadCom1DFA(avalancheDir, cfgMain, runDFAModule=True, cfgFile=com1DFACfgFile)
     for simName in simDF.index:
