+++ benchmarks/avaBowlNullTest/avaBowl_com1DFACfg.ini
@@ -15,6 +15,8 @@ simTypeList = null
 # if secRelArea is True - add secondary release area
 secRelArea = False
 
+dam = False
+
 [INPUT]
 # specify a particular release area scenario, provide name of shapefile with or without extension .shp (optional)
 releaseScenario = release1BL
