+++ benchmarks/avaKotNullTest/avaKot_com1DFACfg.ini
@@ -13,7 +13,8 @@ particleFile = ../benchmarks/avaKotNullTest
 # list of simulations that shall be performed (null, ent, res, entres)
 simTypeList = null
 
+dam = False
+
 [INPUT]
 # specify a particular release area scenario, provide name of shapefile with or without extension .shp (optional)
 releaseScenario = relKot
-
