+++ docs/index.rst
@@ -114,6 +114,7 @@ Theory
     * :doc:`theoryCom1DFA`
     * :doc:`DFAnumerics`
     * :doc:`com1DFAAlgorithm`
+    * :doc:`damToolCom1DFA`
 
 .. toctree::
    :maxdepth: 3
@@ -123,6 +124,7 @@ Theory
    theoryCom1DFA.rst
    DFAnumerics.rst
    com1DFAAlgorithm.rst
+   damToolCom1DFA.rst
 
 .. only:: html
